consolidate gamification
