+++ apps/member-profile/app/routes/_profile.recap.$date.leaderboard.tsx
@@ -1,7 +1,8 @@
 import { json, type LoaderFunctionArgs } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 
-import { getPointsLeaderboard } from '@/member-profile.server';
+import { getPointsLeaderboard } from '@oyster/core/gamification';
+
 import { getDateRange, Recap } from '@/routes/_profile.recap.$date';
 import { Leaderboard } from '@/shared/components/leaderboard';
 import { ensureUserAuthenticated, user } from '@/shared/session.server';
