+++ packages/core/package.json
@@ -22,6 +22,7 @@
     "./employment": "./src/modules/employment/index.ts",
     "./employment.server": "./src/modules/employment/index.server.ts",
     "./events": "./src/modules/event/event.core.ts",
+    "./gamification": "./src/modules/gamification/gamification.core.ts",
     "./github": "./src/modules/github/github.ts",
     "./location": "./src/modules/location/location.core.ts",
     "./location.types": "./src/modules/location/location.types.ts",
