+++ packages/core/src/modules/gamification/gamification.core.ts
@@ -1,8 +1,16 @@
 import { sql } from 'kysely';
+import { match } from 'ts-pattern';
 import { z } from 'zod';
 
 import { db } from '@oyster/db';
 import { Student } from '@oyster/types';
+
+import { GamificationBullJob } from '@/infrastructure/bull/bull.types';
+import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
+import { grantGamificationPoints } from './use-cases/grant-gamification-points';
+import { revokeGamificationPoints } from './use-cases/revoke-gamification-points';
+
+// Queries
 
 type GetPointsLeaderboardOptions = {
   limit: number;
@@ -64,3 +72,53 @@ export async function getPointsLeaderboard({
 
   return leaderboard;
 }
+
+type GetTotalPointsOptions = {
+  occurredAfter?: Date | null;
+};
+
+export async function getTotalPoints(
+  memberId: string,
+  options: GetTotalPointsOptions = {}
+) {
+  const row = await db
+    .selectFrom('completedActivities')
+    .select((eb) => eb.fn.sum('points').as('points'))
+    .where('studentId', '=', memberId)
+    .$if(!!options.occurredAfter, (eb) => {
+      return eb.where('occurredAt', '>=', options.occurredAfter!);
+    })
+    .executeTakeFirstOrThrow();
+
+  const points = Number(row.points);
+
+  return points;
+}
+
+export async function listActivities() {
+  const activities = await db
+    .selectFrom('activities')
+    .select(['description', 'id', 'name', 'period', 'points', 'type'])
+    .where('deletedAt', 'is', null)
+    .orderBy('points', 'asc')
+    .execute();
+
+  return activities;
+}
+
+// Worker
+
+export const gamificationWorker = registerWorker(
+  'gamification',
+  GamificationBullJob,
+  async (job) => {
+    return match(job)
+      .with({ name: 'gamification.activity.completed' }, ({ data }) => {
+        return grantGamificationPoints(data);
+      })
+      .with({ name: 'gamification.activity.completed.undo' }, ({ data }) => {
+        return revokeGamificationPoints(data);
+      })
+      .exhaustive();
+  }
+);
